added function onChange() in storybook
